refact: 구조 변경
