+++ .dockerignore
@@ -1,0 +1,5 @@
+.gitkeep
+.gitattributes
+.gitignore
+docker-compose.yaml
+README.md
