+++ docker-compose.yaml
@@ -2,9 +2,6 @@
 'services':
    'slack_bot':
      'build':
-       'dockerfile': './slack/Dockerfile'
-   'web_server':
-      'build':
-        'dockerfile': './server/Dockerfile'
+       'dockerfile': './Dockerfile'
    'redis':
       'image': 'redis:latest'
