+++ main.py
@@ -1,0 +1,6 @@
+import multiprocessing
+from slack_modules import slack_client
+
+
+if __name__ == "__main__":
+    multiprocessing.Process()
