+++ server_modules/router/main.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+slack_router = APIRouter()
+
+
+@slack_router.post("/incomingWebhook")
+def test_api():
+    return {"code": 200}
