--- slack/Dockerfile
@@ -1,5 +1,0 @@
-FROM python:3.10.10-alpine3.17
-
-COPY . .
-
-CMD ["python","main.py"]
