--- slack/main.py
@@ -1,4 +1,0 @@
-import modules.slack_client
-
-if __name__ == "__main__":
-    print("wait to start")
